Lending
The website is a template for further content filling. A landing page type site. There is an adaptive for all screen resolutions. The accordion functionality is implemented, when you click on a link, it closes one link and opens another Picture with a hover darkening effect. On the Youre tem block. Links to social networks appear in the hover effect. Two sliders. Google maps with address coordinates are connected. There is a burger on the mobile phone permit. The burger will close when you click on the link. Anchor link. The header is fixed and follows the scroll.